% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_split_r.R
\name{choose_split_r}
\alias{choose_split_r}
\title{Choose split direction (recursively)}
\usage{
choose_split_r(row, pretty_tree, model, pred_row)
}
\arguments{
\item{row}{row index of \code{pretty_tree} to determine prediction 
direction}

\item{pred_row}{row contain prediction values to choose split direction}
}
\description{
Choose split direction (recursively)
}
