% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposre_gbm_prediction.R
\name{decompose_gbm_prediction}
\alias{decompose_gbm_prediction}
\title{Decompose gbm prediction into feature contributions}
\usage{
decompose_gbm_prediction(gbm, prediction_row)
}
\arguments{
\item{gbm}{\code{gbm} object to predict with}

\item{prediction_row}{single row \code{data.frame} to predict and the 
decompose into feature contributions}
}
\description{
more detailed explaination of function
}
\details{
based on treeinterpreter https://github.com/andosa/treeinterpreter
}
