% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decision_path.R
\name{get_decision_path}
\alias{get_decision_path}
\title{Determine decision path down tree}
\usage{
get_decision_path(pretty_tree, model, pred_row)
}
\arguments{
\item{pred_row}{row contain prediction values to choose split direction}

\item{row}{row index of \code{pretty_tree} to determine prediction 
direction}
}
\description{
Determine decision path down tree
}
