% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_parents.R
\name{get_node_parents}
\alias{get_node_parents}
\title{Plot tree structure}
\usage{
get_node_parents(pretty_tree)
}
\arguments{
\item{pretty_tree}{\code{data.frame} output from \code{pretty.gbm.tree}
function.}
}
\value{
\code{data.frame} with columns;
\item{\code{parent}}{parent of node}
\item{\code{node}}{node of interest}
}
\description{
Helper function for \code{plot_tree}. Takes node structure output from 
\code{pretty.gbm.tree} which is 1 row per node showing child nodes for that
node and returns the parents of each node.
}
