% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_predictions_xgboost.R
\name{get_node_predictions_xgboost}
\alias{get_node_predictions_xgboost}
\title{Get node predictions for xgboost model}
\usage{
get_node_predictions_xgboost(xgb_model)
}
\arguments{
\item{xgb_model}{\code{xgb.Booster} to extract node predictions from ,}
}
\value{
\code{data.table} with rows representing nodes in trees from the 
model and the following columns;
  \item{row}{data.table row number}
  \item{Tree}{tree index}
  \item{Node}{node index}
  \item{Feature}{variable used for splitting at node or "Leaf" for terminal 
  nodes}
  \item{Split}{split point for node}
  \item{Yes}{tree / node index for 'Yes' child node}
  \item{No}{tree / node index for 'No' child node}
  \item{Missing}{tree / node index for 'Missing' child node}
  \item{Quality}{Gain for node or weight if leaf node}
  \item{Cover}{first order derivative of loss function}
  \item{weight}{node predictions}
  \item{H}{data.table row number}
  \item{G}{Gain for node}
  \item{G_yes}{G value for the 'Yes' child node}
  \item{G_no}{G value for the 'No' child node}
Note the following columns are calculated; row, weight, H, G, G_yes and G_no.
The other columns are output from \code{xgboost::xgb.model.dt.tree}
}
\description{
Get node predictions for xgboost model
}
\examples{
library(xgboost)
data(agaricus.train, package='xgboost')
data(agaricus.test, package='xgboost')
train <- agaricus.train
test <- agaricus.test

bstSparse <- xgboost(data = train$data, 
                     label = train$label, 
                     max.depth = 1, 
                     eta = 1, 
                     nrounds = 2, 
                     objective = "binary:logistic")

get_node_predictions_xgboost(bstSparse)

}
